requirements-spaces.txt
# Core ML libraries
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0

# Web app
streamlit>=1.28.0

# PyTorch ecosystem (CPU only for Spaces)
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0

# AutoML
autogluon.multimodal>=1.0.0

# Hugging Face integration
huggingface_hub>=0.16.0

# Optional: for evaluation and visualization
plotly>=5.15.0
tqdm>=4.65.0 